backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0

sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
asyncmy==0.2.0
pymysql==1.1.0
alembic==1.13.1

pydantic==2.5.0
pydantic-settings==2.1.0

python-dotenv==1.0.0
python-multipart==0.0.6

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

cryptography==41.0.7
bcrypt==4.0.1

httpx>=0.27.0
groq==0.11.0

fastapi-mail==1.4.1

